# docs/swagger.yaml
basePath: /
definitions:
  auth.ErrorResponse:
    properties:
      message:
        example: Invalid param ID
        type: string
    type: object
...
  model.GetMotorcycleBrandsResponse:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
    type: object
  model.GetMotorcycleModelsResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.GetMotorcycleParameterValuesResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.GetMotorcycleParametersResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      values:
        items:
          $ref: '#/definitions/model.GetMotorcycleParameterValuesResponse'
        type: array
    type: object
  model.GetMotorcyclesResponse:
    properties:
      can_look_coordinate:
        type: string
      certificate:
        type: string
      city:
        type: string
      color:
        type: string
      contact_person:
        type: string
      crash:
        type: boolean
      created_at:
        type: string
      date_of_purchase:
        type: string
      description:
        type: string
      email:
        type: string
      engine:
        type: integer
      fuel_type:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      moto_brand:
        type: string
      moto_category:
        type: string
      moto_model:
        type: string
      my_car:
        type: boolean
      not_cleared:
        type: boolean
      number_of_cycles:
        type: integer
      odometer:
        type: integer
      only_chat:
        type: boolean
      owner:
        $ref: '#/definitions/model.MotorcycleOwner'
      owners:
        type: integer
      parameters:
        items:
          $ref: '#/definitions/model.MotorcycleParameter'
        type: array
      phone_number:
        type: string
      power:
        type: integer
      price:
        type: integer
      price_type:
        type: string
      protect_spam:
        type: boolean
      ptc:
        type: boolean
      refuse_dealers_calls:
        type: boolean
      status:
        type: string
      updated_at:
        type: string
      verified_buyers:
        type: boolean
      videos:
        items:
          type: string
        type: array
      vin_code:
        type: string
      warranty_date:
        type: string
      year:
        type: integer
    type: object
...
  model.MotorcycleOwner:
    properties:
      avatar:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  model.MotorcycleParameter:
    properties:
      parameter:
        type: string
      parameter_id:
        type: integer
      parameter_value:
        type: string
      parameter_value_id:
        type: integer
    type: object
...
  model.ResultMessage:
    properties:
      en:
        type: string
      ru:
        type: string
      tk:
        type: string
    type: object
  model.Success:
    properties:
      message:
        type: string
    type: object
  model.SuccessWithId:
    properties:
      id:
        type: integer
      message:
        type: string
    type: object
...
host: 84.200.87.48
info:
  contact: {}
  description: Project Description
  title: Project name
  version: "1.0"
paths:
...
  /motorcycles:
    get:
      consumes:
      - application/json
      description: Get motorcycles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GetMotorcyclesResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Get motorcycles
      tags:
      - motorcycles
    post:
      consumes:
      - application/json
      description: Create motorcycle
      parameters:
      - description: Motorcycle
        in: body
        name: motorcycle
        required: true
        schema:
          $ref: '#/definitions/model.CreateMotorcycleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessWithId'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Create motorcycle
      tags:
      - motorcycles
  /motorcycles/{id}:
    delete:
      description: Deletes a motorcycle and its associated files
      parameters:
      - description: Motorcycle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Delete motorcycle
      tags:
      - motorcycle
    get:
      description: Returns a motorcycle by its ID
      parameters:
      - description: Motorcycle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetMotorcyclesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Get motorcycle by ID
      tags:
      - motorcycle
  /motorcycles/{id}/buy:
    post:
      description: Returns a status response message
      parameters:
      - description: Motorcycle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Buy motorcycle
      tags:
      - motorcycle
  /motorcycles/{id}/dont-sell:
    post:
      description: Updates motorcycle status to not for sale
      parameters:
      - description: Motorcycle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Set motorcycle as not for sale
      tags:
      - motorcycle
  /motorcycles/{id}/edit:
    get:
      description: Returns a motorcycle by its ID for editing
      parameters:
      - description: Motorcycle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetMotorcyclesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Get Edit motorcycle by ID
      tags:
      - motorcycle
  /motorcycles/{id}/sell:
    post:
      description: Updates motorcycle status to for sale
      parameters:
      - description: Motorcycle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Set motorcycle for sale
      tags:
      - motorcycle
  /motorcycles/{motorcycle_id}/images:
    post:
      consumes:
      - multipart/form-data
      description: Uploads images for a motorcycle (max 10 files)
      parameters:
      - description: Motorcycle ID
        in: path
        name: motorcycle_id
        required: true
        type: integer
      - description: Motorcycle images (max 10)
        in: formData
        name: images
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Upload motorcycle images
      tags:
      - motorcycle
  /motorcycles/{motorcycle_id}/images/{image_id}:
    delete:
      consumes:
      - application/json
      description: Deletes an image from a motorcycle
      parameters:
      - description: Motorcycle ID
        in: path
        name: motorcycle_id
        required: true
        type: integer
      - description: Image ID
        in: path
        name: image_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Delete motorcycle image
      tags:
      - motorcycle
  /motorcycles/{motorcycle_id}/videos:
    post:
      consumes:
      - multipart/form-data
      description: Uploads videos for a motorcycle (max 1 files)
      parameters:
      - description: Motorcycle ID
        in: path
        name: motorcycle_id
        required: true
        type: integer
      - description: Motorcycle videos (max 10)
        in: formData
        name: videos
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Upload motorcycle videos
      tags:
      - motorcycle
  /motorcycles/{motorcycle_id}/videos/{video_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a video from a motorcycle
      parameters:
      - description: Motorcycle ID
        in: path
        name: motorcycle_id
        required: true
        type: integer
      - description: Video ID
        in: path
        name: video_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Delete motorcycle video
      tags:
      - motorcycle
  /motorcycles/categories/{category_id}/brands:
    get:
      consumes:
      - application/json
      description: Get motorcycle brands
      parameters:
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GetMotorcycleBrandsResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      summary: Get motorcycle brands
      tags:
      - motorcycles
  /motorcycles/categories/{category_id}/brands/{brand_id}/models:
    get:
      consumes:
      - application/json
      description: Get motorcycle models by brand ID
      parameters:
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      - description: Brand ID
        in: path
        name: brand_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GetMotorcycleModelsResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      summary: Get motorcycle models by brand ID
      tags:
      - motorcycles
  /motorcycles/categories/{category_id}/parameters:
    get:
      consumes:
      - application/json
      description: Get motorcycle parameters
      parameters:
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GetMotorcycleParametersResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      summary: Get motorcycle parameters
      tags:
      - motorcycles
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"